+++ src/libexpr/eval-gc.hh
@@ -1,0 +1,16 @@
+#pragma once
+///@file
+
+namespace nix {
+
+/**
+ * Initialise the Boehm GC, if applicable.
+ */
+void initGC();
+
+/**
+ * Make sure `initGC` has already been called.
+ */
+void assertGCInitialized();
+
+}
